Fix FFmpeg error
