+++ tests/test_ffmpeg_eager.py
@@ -124,3 +124,10 @@ def test_ffmpeg_decode_video():
   video = ffmpeg_io.decode_video(content, 0)
   assert video.shape == [166, 320, 560, 3]
   assert video.dtype == tf.uint8
+
+def test_video_predict():
+  model = tf.keras.applications.resnet50.ResNet50(weights='imagenet')
+  x = ffmpeg_io.VideoDataset(video_path).batch(1).map(lambda x: tf.keras.applications.resnet50.preprocess_input(tf.image.resize(x, (224, 224))))
+  y = model.predict(x)
+  p = tf.keras.applications.resnet50.decode_predictions(y, top=1)
+  assert len(p) == 166
